formated int output

+++ _printf.c
@@ -31,7 +31,7 @@ int _printf(const char *format, ...)
 					num_of_char += fprintf(stdout, "%s", va_arg(the_argmnts, char *));
 					break;
 				case 'd':
-					num_of_char += fprintf(stdout, "%d", va_arg(the_argmnts, double));
+					num_of_char += fprintf(stdout, "%d", va_arg(the_argmnts, int));
 					break;
 				case 'i':
 					num_of_char += fprintf(stdout, "%d", va_arg(the_argmnts, int));
